Rename file hot-news-banner.vue to HotNewsBanner.vue
