+++ components/HotNewsBanner.vue
